+++ src/test/scala/edu/ie3/simona/sim/setup/ExtSimSetupDataSpec.scala
@@ -15,6 +15,7 @@ import edu.ie3.simona.api.data.connection.{
   ExtResultDataConnection,
 }
 import edu.ie3.simona.ontology.messages.ServiceMessage
+import edu.ie3.simona.service.em.ExtEmDataService
 import edu.ie3.simona.service.ev.ExtEvDataService
 import edu.ie3.simona.service.primary.PrimaryServiceProxy
 import edu.ie3.simona.test.common.UnitSpec
@@ -46,8 +47,9 @@ class ExtSimSetupDataSpec extends ScalaTestWithActorTestKit with UnitSpec {
       )
 
       updated.extSimAdapters shouldBe empty
-      updated.extPrimaryDataServices shouldBe Seq((connection, primaryRef))
-      updated.extDataServices shouldBe empty
+      updated.primaryDataServices shouldBe Seq((connection, primaryRef))
+      updated.emDataService shouldBe None
+      updated.evDataService shouldBe None
       updated.extResultListeners shouldBe empty
     }
 
@@ -67,11 +69,12 @@ class ExtSimSetupDataSpec extends ScalaTestWithActorTestKit with UnitSpec {
         .update(connection2, primaryRef2)
 
       updated.extSimAdapters shouldBe empty
-      updated.extPrimaryDataServices shouldBe Seq(
+      updated.primaryDataServices shouldBe Seq(
         (connection1, primaryRef1),
         (connection2, primaryRef2),
       )
-      updated.extDataServices shouldBe empty
+      updated.emDataService shouldBe None
+      updated.evDataService shouldBe None
       updated.extResultListeners shouldBe empty
     }
 
@@ -86,26 +89,26 @@ class ExtSimSetupDataSpec extends ScalaTestWithActorTestKit with UnitSpec {
       val evRef = TestProbe[ExtEvDataService.Message]("ev_service").ref
 
       val emConnection = new ExtEmDataConnection(emptyUuidList, EmMode.BASE)
-      val emRef = TestProbe[ServiceMessage]("em_service").ref
+      val emRef = TestProbe[ExtEmDataService.Message]("em_service").ref
 
       val cases = Table(
         ("connection", "serviceRef", "expected"),
         (
           primaryConnection,
           primaryRef,
-          extSimSetupData.copy(extPrimaryDataServices =
+          extSimSetupData.copy(primaryDataServices =
             Seq((primaryConnection, primaryRef))
           ),
         ),
         (
           evConnection,
           evRef,
-          extSimSetupData.copy(extDataServices = Seq((evConnection, evRef))),
+          extSimSetupData.copy(evDataService = Some(evRef)),
         ),
         (
           emConnection,
           emRef,
-          extSimSetupData.copy(extDataServices = Seq((emConnection, emRef))),
+          extSimSetupData.copy(emDataService = Some(emRef)),
         ),
       )
 
@@ -113,9 +116,10 @@ class ExtSimSetupDataSpec extends ScalaTestWithActorTestKit with UnitSpec {
         val updated = extSimSetupData.update(connection, serviceRef)
 
         updated.extSimAdapters shouldBe expected.extSimAdapters
-        updated.extPrimaryDataServices shouldBe expected.extPrimaryDataServices
-        updated.extDataServices shouldBe expected.extDataServices
-        updated.extResultListeners shouldBe expected.extResultListeners
+        updated.primaryDataServices shouldBe expected.primaryDataServices
+        updated.emDataService shouldBe expected.emDataService
+        updated.evDataService shouldBe expected.evDataService
+        updated.extResultListeners shouldBe empty
       }
     }
 
@@ -128,12 +132,13 @@ class ExtSimSetupDataSpec extends ScalaTestWithActorTestKit with UnitSpec {
       val updated = extSimSetupData.update(resultConnection, resultRef)
 
       updated.extSimAdapters shouldBe empty
-      updated.extPrimaryDataServices shouldBe empty
-      updated.extDataServices shouldBe empty
-      updated.extResultListeners shouldBe Seq((resultConnection, resultRef))
+      updated.primaryDataServices shouldBe empty
+      updated.emDataService shouldBe None
+      updated.evDataService shouldBe None
+      updated.extResultListeners shouldBe Seq(resultRef)
     }
 
-    "be updated with multiple different connections" in {
+    "be updated with multiple different connections correctly" in {
       val extSimSetupData = ExtSimSetupData.apply
 
       val primaryConnection = new ExtPrimaryDataConnection(emptyMapInput)
@@ -144,29 +149,27 @@ class ExtSimSetupDataSpec extends ScalaTestWithActorTestKit with UnitSpec {
       val evRef = TestProbe[ExtEvDataService.Message]("ev_service").ref
 
       val emConnection = new ExtEmDataConnection(emptyUuidList, EmMode.BASE)
-      val emRef = TestProbe[ServiceMessage]("em_service").ref
+      val emRef = TestProbe[ExtEmDataService.Message]("em_service").ref
 
       val resultConnection = new ExtResultDataConnection(emptyUuidList)
       val resultRef = TestProbe[ServiceMessage]("result_service").ref
 
       val updated = extSimSetupData
         .update(primaryConnection, primaryRef)
-        .update(evConnection, evRef)
         .update(emConnection, emRef)
+        .update(evConnection, evRef)
         .update(resultConnection, resultRef)
 
       updated.extSimAdapters shouldBe empty
-      updated.extPrimaryDataServices shouldBe Seq(
+      updated.primaryDataServices shouldBe Seq(
         (
           primaryConnection,
           primaryRef,
         )
       )
-      updated.extDataServices shouldBe Seq(
-        (evConnection, evRef),
-        (emConnection, emRef),
-      )
-      updated.extResultListeners shouldBe Seq((resultConnection, resultRef))
+      updated.emDataService shouldBe Some(emRef)
+      updated.evDataService shouldBe Some(evRef)
+      updated.extResultListeners shouldBe Seq(resultRef)
     }
 
     "return evDataService correctly" in {
@@ -174,17 +177,12 @@ class ExtSimSetupDataSpec extends ScalaTestWithActorTestKit with UnitSpec {
       val evRef = TestProbe[ExtEvDataService.Message]("ev_service").ref
 
       val cases = Table(
-        ("extSimSetupData", "expectedConnection", "expectedService"),
-        (
-          ExtSimSetupData.apply.update(evConnection, evRef),
-          Some(evConnection),
-          Some(evRef),
-        ),
-        (ExtSimSetupData.apply, None, None),
+        ("extSimSetupData", "expectedService"),
+        (ExtSimSetupData.apply.update(evConnection, evRef), Some(evRef)),
+        (ExtSimSetupData.apply, None),
       )
 
-      forAll(cases) { (extSimSetupData, expectedConnection, expectedService) =>
-        extSimSetupData.evDataConnection shouldBe expectedConnection
+      forAll(cases) { (extSimSetupData, expectedService) =>
         extSimSetupData.evDataService shouldBe expectedService
       }
     }
@@ -204,11 +202,9 @@ class ExtSimSetupDataSpec extends ScalaTestWithActorTestKit with UnitSpec {
       )
 
       forAll(cases) { (extSimSetupData, expectedConnection, expectedService) =>
-        extSimSetupData.emDataConnection shouldBe expectedConnection
         extSimSetupData.emDataService shouldBe expectedService
       }
     }
-
   }
 
 }